+++ src/Core/VRChatContentPublisher.TelemetryCore/Extensions/TracerProviderBuilderExtension.cs
@@ -1,5 +1,6 @@
 ﻿using OpenTelemetry.Trace;
 using VRChatContentPublisher.TelemetryCore.Masking.OpenTelemetry;
+using VRChatContentPublisher.TelemetryCore.OpenTelemetryProcessor;
 using VRChatContentPublisher.TelemetryCore.TelemetryToggle;
 
 namespace VRChatContentPublisher.TelemetryCore.Extensions;
@@ -13,7 +14,8 @@ public static class TracerProviderBuilderExtension
         builder.SetSampler(new AppOpenTelemetryToggleSampler());
 
         builder.AddProcessor(new CustomTagsProcessor([
-            new CustomTagsProcessorTag("app.lifetime_session_id", appSessionLifetimeId)
+            new CustomTagsProcessorTag("app.lifetime_session_id", appSessionLifetimeId),
+            new CustomTagsProcessorTag("user.id", InstallationIdProvider.GetInstallationId())
         ]));
 
         builder.AddProcessor(new EnvironmentTagProcessor());